Rename module form data_base.py to database.py
#357
