+++ modules/layer/__init__.py
@@ -1,4 +1,4 @@
 from .add_layer import LayerAdder
 from .geo_service import GeoServiceManager
-from .data_base import DataBaseManager
+from .database import DataBaseManager
 from .limitations_checker import LimitationsChecker
